module-api/src/main/java/com/yapp/sharefood/food/service/FoodService.java
package com.yapp.sharefood.food.service;

import com.yapp.sharefood.food.domain.Food;
import com.yapp.sharefood.food.dto.request.FoodCreationRequest;
import com.yapp.sharefood.food.repository.FoodRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@RequiredArgsConstructor
@Transactional(readOnly = true)
public class FoodService {

    private final FoodRepository foodRepository;

    @Transactional
    public Long saveFood(FoodCreationRequest foodCreationRequest) {
        Food food = Food.builder()
                .foodTitle(foodCreationRequest.getTitle())
                .foodStatus(foodCreationRequest.getFoodStatus())
                .price(foodCreationRequest.getPrice())
                .reviewMsg(foodCreationRequest.getReviewMsg())
                .build();

        Food save = foodRepository.save(food);

        return save.getId();
    }
}
